refactor: share desktop bridge contract types
