+++ packages/app/src/app.tsx
@@ -44,12 +44,12 @@ import { ServerConnection, ServerProvider, serverName, useServer } from "@/conte
 import { SettingsProvider } from "@/context/settings"
 import { TerminalProvider } from "@/context/terminal"
 import { AppStartupPending } from "@/components/app-startup-pending"
-import { AboutModal, type AboutInfo } from "@/components/about-modal"
+import { AboutModal } from "@/components/about-modal"
+import type { AboutInfo, RendererDiagnosticInput, RendererDiagnosticsExportResult } from "@/desktop-api-contract"
 import DirectoryLayout from "@/pages/directory-layout"
 import Layout from "@/pages/layout"
 import { ErrorPage } from "./pages/error"
 import { buildDesktopContext, desktopWindowTitle, type DesktopContext } from "./utils/desktop-context"
-import type { RendererDiagnosticInput, RendererDiagnosticsExportResult } from "@/context/platform"
 import { useCheckServerHealth } from "./utils/server-health"
 import { base64Encode } from "@opencode-ai/util/encode"
