+++ packages/app/src/desktop-api.ts
@@ -1,12 +1,16 @@
 export { buildDesktopContext, desktopWindowTitle, type DesktopContext } from "./utils/desktop-context"
 export type {
+  AboutInfo,
+  RendererDiagnosticInput,
+  RendererDiagnosticsExportResult,
+  RendererErrorDetails,
+  ReportProblemInput,
+  ReportProblemResult,
+  UpdateInfo,
+} from "./desktop-api-contract"
+export type {
   BrowserBridge,
   BrowserState,
   BrowserViewLayout,
   BrowserViewRect,
-  RendererDiagnosticInput,
-  RendererDiagnosticsExportResult,
-  ReportProblemInput,
-  ReportProblemResult,
-  UpdateInfo,
 } from "./context/platform"
